refator(exporter) Shopware6 command and messenger configuration
