+++ module/exporter-shopware-6/src/Domain/Command/Export/CategoryExportCommand.php
@@ -13,7 +13,7 @@ use Ergonode\SharedKernel\Domain\Aggregate\CategoryId;
 use Ergonode\SharedKernel\Domain\Aggregate\ExportId;
 use JMS\Serializer\Annotation as JMS;
 
-class CategoryShopware6ExportCommand implements ExporterCommandInterface
+class CategoryExportCommand implements ExporterCommandInterface
 {
     /**
      * @JMS\Type("Ergonode\SharedKernel\Domain\Aggregate\ExportId")
