+++ module/exporter-shopware-6/src/Domain/Command/Export/CategoryRemoveExportCommand.php
@@ -13,7 +13,7 @@ use Ergonode\SharedKernel\Domain\Aggregate\CategoryId;
 use Ergonode\SharedKernel\Domain\Aggregate\ExportId;
 use JMS\Serializer\Annotation as JMS;
 
-class CategoryRemoveShopware6ExportCommand implements ExporterCommandInterface
+class CategoryRemoveExportCommand implements ExporterCommandInterface
 {
     /**
      * @JMS\Type("Ergonode\SharedKernel\Domain\Aggregate\ExportId")
