+++ module/exporter-shopware-6/src/Domain/Command/Export/CustomFieldExportCommand.php
@@ -13,7 +13,7 @@ use Ergonode\SharedKernel\Domain\Aggregate\AttributeId;
 use Ergonode\SharedKernel\Domain\Aggregate\ExportId;
 use JMS\Serializer\Annotation as JMS;
 
-class PropertyGroupShopware6ExportCommand implements ExporterCommandInterface
+class CustomFieldExportCommand implements ExporterCommandInterface
 {
     /**
      * @JMS\Type("Ergonode\SharedKernel\Domain\Aggregate\ExportId")
