+++ module/exporter-shopware-6/src/Infrastructure/Handler/Export/CategoryExportCommandHandler.php
@@ -8,46 +8,47 @@ declare(strict_types=1);
 
 namespace Ergonode\ExporterShopware6\Infrastructure\Handler\Export;
 
-use Ergonode\Attribute\Domain\Entity\AbstractAttribute;
-use Ergonode\Attribute\Domain\Repository\AttributeRepositoryInterface;
+use Ergonode\Category\Domain\Entity\AbstractCategory;
+use Ergonode\Category\Domain\Repository\CategoryRepositoryInterface;
 use Ergonode\Channel\Domain\Repository\ChannelRepositoryInterface;
 use Ergonode\Exporter\Domain\Entity\Export;
 use Ergonode\Exporter\Domain\Repository\ExportRepositoryInterface;
-use Ergonode\ExporterShopware6\Domain\Command\Export\PropertyGroupShopware6ExportCommand;
+use Ergonode\ExporterShopware6\Domain\Command\Export\CategoryExportCommand;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
-use Ergonode\ExporterShopware6\Infrastructure\Processor\Process\PropertyGroupShopware6ExportProcess;
+use Ergonode\ExporterShopware6\Infrastructure\Processor\Process\CategoryShopware6ExportProcess;
 use Webmozart\Assert\Assert;
 
-class PropertyGroupShopware6ExportCommandHandler
+class CategoryExportCommandHandler
 {
     private ExportRepositoryInterface $exportRepository;
 
     private ChannelRepositoryInterface $channelRepository;
 
-    private AttributeRepositoryInterface $attributeRepository;
-    private PropertyGroupShopware6ExportProcess $process;
+    private CategoryRepositoryInterface $categoryRepository;
+
+    private CategoryShopware6ExportProcess $process;
 
     public function __construct(
         ExportRepositoryInterface $exportRepository,
         ChannelRepositoryInterface $channelRepository,
-        AttributeRepositoryInterface $attributeRepository,
-        PropertyGroupShopware6ExportProcess $process
+        CategoryRepositoryInterface $categoryRepository,
+        CategoryShopware6ExportProcess $process
     ) {
         $this->exportRepository = $exportRepository;
         $this->channelRepository = $channelRepository;
-        $this->attributeRepository = $attributeRepository;
+        $this->categoryRepository = $categoryRepository;
         $this->process = $process;
     }
 
-    public function __invoke(PropertyGroupShopware6ExportCommand $command): void
+    public function __invoke(CategoryExportCommand $command): void
     {
         $export = $this->exportRepository->load($command->getExportId());
         Assert::isInstanceOf($export, Export::class);
         $channel = $this->channelRepository->load($export->getChannelId());
         Assert::isInstanceOf($channel, Shopware6Channel::class);
-        $attribute = $this->attributeRepository->load($command->getAttributeId());
-        Assert::isInstanceOf($attribute, AbstractAttribute::class);
+        $category = $this->categoryRepository->load($command->getCategoryId());
+        Assert::isInstanceOf($category, AbstractCategory::class);
 
-        $this->process->process($export, $channel, $attribute);
+        $this->process->process($export, $channel, $category, $command->getParentCategoryId());
     }
 }
