+++ module/exporter-shopware-6/src/Infrastructure/Handler/Export/ProductCrossSellingExportCommandHandler.php
@@ -8,47 +8,47 @@ declare(strict_types=1);
 
 namespace Ergonode\ExporterShopware6\Infrastructure\Handler\Export;
 
-use Ergonode\Category\Domain\Entity\AbstractCategory;
-use Ergonode\Category\Domain\Repository\CategoryRepositoryInterface;
 use Ergonode\Channel\Domain\Repository\ChannelRepositoryInterface;
 use Ergonode\Exporter\Domain\Entity\Export;
 use Ergonode\Exporter\Domain\Repository\ExportRepositoryInterface;
-use Ergonode\ExporterShopware6\Domain\Command\Export\CategoryShopware6ExportCommand;
+use Ergonode\ExporterShopware6\Domain\Command\Export\ProductCrossSellingExportCommand;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
-use Ergonode\ExporterShopware6\Infrastructure\Processor\Process\CategoryShopware6ExportProcess;
+use Ergonode\ExporterShopware6\Infrastructure\Processor\Process\ProductCrossSellingExportProcess;
+use Ergonode\ProductCollection\Domain\Entity\ProductCollection;
+use Ergonode\ProductCollection\Domain\Repository\ProductCollectionRepositoryInterface;
 use Webmozart\Assert\Assert;
 
-class CategoryShopware6ExportCommandHandler
+class ProductCrossSellingExportCommandHandler
 {
     private ExportRepositoryInterface $exportRepository;
 
     private ChannelRepositoryInterface $channelRepository;
 
-    private CategoryRepositoryInterface $categoryRepository;
+    private ProductCollectionRepositoryInterface $productCollectionRepository;
 
-    private CategoryShopware6ExportProcess $process;
+    private ProductCrossSellingExportProcess $process;
 
     public function __construct(
         ExportRepositoryInterface $exportRepository,
         ChannelRepositoryInterface $channelRepository,
-        CategoryRepositoryInterface $categoryRepository,
-        CategoryShopware6ExportProcess $process
+        ProductCollectionRepositoryInterface $productCollectionRepository,
+        ProductCrossSellingExportProcess $process
     ) {
         $this->exportRepository = $exportRepository;
         $this->channelRepository = $channelRepository;
-        $this->categoryRepository = $categoryRepository;
+        $this->productCollectionRepository = $productCollectionRepository;
         $this->process = $process;
     }
 
-    public function __invoke(CategoryShopware6ExportCommand $command): void
+    public function __invoke(ProductCrossSellingExportCommand $command): void
     {
         $export = $this->exportRepository->load($command->getExportId());
         Assert::isInstanceOf($export, Export::class);
         $channel = $this->channelRepository->load($export->getChannelId());
         Assert::isInstanceOf($channel, Shopware6Channel::class);
-        $category = $this->categoryRepository->load($command->getCategoryId());
-        Assert::isInstanceOf($category, AbstractCategory::class);
+        $productCollection = $this->productCollectionRepository->load($command->getProductCollectionId());
+        Assert::isInstanceOf($productCollection, ProductCollection::class);
 
-        $this->process->process($export, $channel, $category, $command->getParentCategoryId());
+        $this->process->process($export, $channel, $productCollection);
     }
 }
