+++ module/exporter-shopware-6/src/Infrastructure/Handler/Export/PropertyGroupExportCommandHandler.php
@@ -8,47 +8,46 @@ declare(strict_types=1);
 
 namespace Ergonode\ExporterShopware6\Infrastructure\Handler\Export;
 
+use Ergonode\Attribute\Domain\Entity\AbstractAttribute;
+use Ergonode\Attribute\Domain\Repository\AttributeRepositoryInterface;
 use Ergonode\Channel\Domain\Repository\ChannelRepositoryInterface;
 use Ergonode\Exporter\Domain\Entity\Export;
 use Ergonode\Exporter\Domain\Repository\ExportRepositoryInterface;
-use Ergonode\ExporterShopware6\Domain\Command\Export\ProductCrossSellingShopware6ExportCommand;
+use Ergonode\ExporterShopware6\Domain\Command\Export\PropertyGroupExportCommand;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
-use Ergonode\ExporterShopware6\Infrastructure\Processor\Process\ProductCrossSellingExportProcess;
-use Ergonode\ProductCollection\Domain\Entity\ProductCollection;
-use Ergonode\ProductCollection\Domain\Repository\ProductCollectionRepositoryInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Processor\Process\PropertyGroupShopware6ExportProcess;
 use Webmozart\Assert\Assert;
 
-class ProductCrossSellingShopware6ExportCommandHandler
+class PropertyGroupExportCommandHandler
 {
     private ExportRepositoryInterface $exportRepository;
 
     private ChannelRepositoryInterface $channelRepository;
 
-    private ProductCollectionRepositoryInterface $productCollectionRepository;
-
-    private ProductCrossSellingExportProcess $process;
+    private AttributeRepositoryInterface $attributeRepository;
+    private PropertyGroupShopware6ExportProcess $process;
 
     public function __construct(
         ExportRepositoryInterface $exportRepository,
         ChannelRepositoryInterface $channelRepository,
-        ProductCollectionRepositoryInterface $productCollectionRepository,
-        ProductCrossSellingExportProcess $process
+        AttributeRepositoryInterface $attributeRepository,
+        PropertyGroupShopware6ExportProcess $process
     ) {
         $this->exportRepository = $exportRepository;
         $this->channelRepository = $channelRepository;
-        $this->productCollectionRepository = $productCollectionRepository;
+        $this->attributeRepository = $attributeRepository;
         $this->process = $process;
     }
 
-    public function __invoke(ProductCrossSellingShopware6ExportCommand $command): void
+    public function __invoke(PropertyGroupExportCommand $command): void
     {
         $export = $this->exportRepository->load($command->getExportId());
         Assert::isInstanceOf($export, Export::class);
         $channel = $this->channelRepository->load($export->getChannelId());
         Assert::isInstanceOf($channel, Shopware6Channel::class);
-        $productCollection = $this->productCollectionRepository->load($command->getProductCollectionId());
-        Assert::isInstanceOf($productCollection, ProductCollection::class);
+        $attribute = $this->attributeRepository->load($command->getAttributeId());
+        Assert::isInstanceOf($attribute, AbstractAttribute::class);
 
-        $this->process->process($export, $channel, $productCollection);
+        $this->process->process($export, $channel, $attribute);
     }
 }
