+++ module/exporter-shopware-6/src/Infrastructure/Processor/Step/Shopware6CategoryStep.php
@@ -12,7 +12,7 @@ use Ergonode\Category\Domain\Entity\CategoryTree;
 use Ergonode\Category\Domain\Repository\TreeRepositoryInterface;
 use Ergonode\Category\Domain\ValueObject\Node;
 use Ergonode\EventSourcing\Infrastructure\Bus\CommandBusInterface;
-use Ergonode\ExporterShopware6\Domain\Command\Export\CategoryShopware6ExportCommand;
+use Ergonode\ExporterShopware6\Domain\Command\Export\CategoryExportCommand;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
 use Ergonode\ExporterShopware6\Infrastructure\Processor\Shopware6ExportStepProcessInterface;
 use Ergonode\SharedKernel\Domain\Aggregate\CategoryId;
@@ -46,7 +46,7 @@ class Shopware6CategoryStep implements Shopware6ExportStepProcessInterface
 
     private function buildStep(ExportId $exportId, Node $node, CategoryId $parentId = null): void
     {
-        $processCommand = new CategoryShopware6ExportCommand($exportId, $node->getCategoryId(), $parentId);
+        $processCommand = new CategoryExportCommand($exportId, $node->getCategoryId(), $parentId);
         $this->commandBus->dispatch($processCommand, true);
 
         $newParent = $node->getCategoryId();
