+++ module/exporter-shopware-6/src/Resources/config/messenger.yaml
@@ -1,8 +1,10 @@
 framework:
     messenger:
         routing:
-            Ergonode\ExporterShopware6\Domain\Command\Export\ProductShopware6ExportCommand: '%ergonode.exporter_shopware_6.messenger_transport_name%'
-            Ergonode\ExporterShopware6\Domain\Command\Export\PropertyGroupShopware6ExportCommand: '%ergonode.exporter_shopware_6.messenger_transport_name%'
-            Ergonode\ExporterShopware6\Domain\Command\Export\CustomFieldShopware6ExportCommand: '%ergonode.exporter_shopware_6.messenger_transport_name%'
-            Ergonode\ExporterShopware6\Domain\Command\Export\CategoryShopware6ExportCommand: '%ergonode.exporter_shopware_6.messenger_transport_name%'
+            Ergonode\ExporterShopware6\Domain\Command\Export\ProductExportCommand: '%ergonode.exporter_shopware_6.messenger_transport_name%'
+            Ergonode\ExporterShopware6\Domain\Command\Export\PropertyGroupExportCommand: '%ergonode.exporter_shopware_6.messenger_transport_name%'
+            Ergonode\ExporterShopware6\Domain\Command\Export\CustomFieldExportCommand: '%ergonode.exporter_shopware_6.messenger_transport_name%'
+            Ergonode\ExporterShopware6\Domain\Command\Export\CategoryExportCommand: '%ergonode.exporter_shopware_6.messenger_transport_name%'
+            Ergonode\ExporterShopware6\Domain\Command\Export\CategoryRemoveExportCommand: '%ergonode.exporter_shopware_6.messenger_transport_name%'
+            Ergonode\ExporterShopware6\Domain\Command\Export\ProductCrossSellingExportCommand: '%ergonode.exporter_shopware_6.messenger_transport_name%'
             Ergonode\ExporterShopware6\Domain\Command\Export\EndShopware6ExportCommand: '%ergonode.exporter_shopware_6.messenger_transport_name%'
