+++ module/exporter-shopware-6/tests/Infrastructure/Handler/Export/CategoryExportCommandHandlerTest.php
@@ -8,19 +8,19 @@ declare(strict_types=1);
 
 namespace Ergonode\ExporterShopware6\Tests\Infrastructure\Handler\Export;
 
-use Ergonode\Attribute\Domain\Entity\AbstractAttribute;
-use Ergonode\Attribute\Domain\Repository\AttributeRepositoryInterface;
+use Ergonode\Category\Domain\Entity\AbstractCategory;
+use Ergonode\Category\Domain\Repository\CategoryRepositoryInterface;
 use Ergonode\Channel\Domain\Repository\ChannelRepositoryInterface;
 use Ergonode\Exporter\Domain\Entity\Export;
 use Ergonode\Exporter\Domain\Repository\ExportRepositoryInterface;
-use Ergonode\ExporterShopware6\Domain\Command\Export\CustomFieldShopware6ExportCommand;
+use Ergonode\ExporterShopware6\Domain\Command\Export\CategoryExportCommand;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
-use Ergonode\ExporterShopware6\Infrastructure\Handler\Export\CustomFieldShopware6ExportCommandHandler;
-use Ergonode\ExporterShopware6\Infrastructure\Processor\Process\CustomFiledShopware6ExportProcess;
+use Ergonode\ExporterShopware6\Infrastructure\Handler\Export\CategoryExportCommandHandler;
+use Ergonode\ExporterShopware6\Infrastructure\Processor\Process\CategoryShopware6ExportProcess;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
-class CustomFieldShopware6ExportCommandHandlerTest extends TestCase
+class CategoryExportCommandHandlerTest extends TestCase
 {
     /**
      * @var ExportRepositoryInterface|MockObject
@@ -33,14 +33,14 @@ class CustomFieldShopware6ExportCommandHandlerTest extends TestCase
     private ChannelRepositoryInterface $channelRepository;
 
     /**
-     * @var AttributeRepositoryInterface|MockObject
+     * @var CategoryRepositoryInterface|MockObject
      */
-    private AttributeRepositoryInterface $attributeRepository;
+    private CategoryRepositoryInterface $categoryRepository;
 
     /**
-     * @var CustomFiledShopware6ExportProcess|MockObject
+     * @var CategoryShopware6ExportProcess|MockObject
      */
-    private CustomFiledShopware6ExportProcess $process;
+    private CategoryShopware6ExportProcess $process;
 
     protected function setUp(): void
     {
@@ -54,23 +54,23 @@ class CustomFieldShopware6ExportCommandHandlerTest extends TestCase
             ->willReturn($this->createMock(Shopware6Channel::class));
         $this->channelRepository->expects(self::once())->method('load');
 
-        $this->attributeRepository = $this->createMock(AttributeRepositoryInterface::class);
-        $this->attributeRepository->method('load')
-            ->willReturn($this->createMock(AbstractAttribute::class));
-        $this->attributeRepository->expects(self::once())->method('load');
+        $this->categoryRepository = $this->createMock(CategoryRepositoryInterface::class);
+        $this->categoryRepository->method('load')
+            ->willReturn($this->createMock(AbstractCategory::class));
+        $this->categoryRepository->expects(self::once())->method('load');
 
-        $this->process = $this->createMock(CustomFiledShopware6ExportProcess::class);
+        $this->process = $this->createMock(CategoryShopware6ExportProcess::class);
         $this->process->expects(self::once())->method('process');
     }
 
     public function testHandling(): void
     {
-        $command = $this->createMock(CustomFieldShopware6ExportCommand::class);
+        $command = $this->createMock(CategoryExportCommand::class);
 
-        $handler = new CustomFieldShopware6ExportCommandHandler(
+        $handler = new CategoryExportCommandHandler(
             $this->exportRepository,
             $this->channelRepository,
-            $this->attributeRepository,
+            $this->categoryRepository,
             $this->process
         );
         $handler->__invoke($command);
