+++ module/exporter-shopware-6/tests/Infrastructure/Handler/Export/CustomFieldExportCommandHandlerTest.php
@@ -13,14 +13,14 @@ use Ergonode\Attribute\Domain\Repository\AttributeRepositoryInterface;
 use Ergonode\Channel\Domain\Repository\ChannelRepositoryInterface;
 use Ergonode\Exporter\Domain\Entity\Export;
 use Ergonode\Exporter\Domain\Repository\ExportRepositoryInterface;
-use Ergonode\ExporterShopware6\Domain\Command\Export\PropertyGroupShopware6ExportCommand;
+use Ergonode\ExporterShopware6\Domain\Command\Export\CustomFieldExportCommand;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
-use Ergonode\ExporterShopware6\Infrastructure\Handler\Export\PropertyGroupShopware6ExportCommandHandler;
-use Ergonode\ExporterShopware6\Infrastructure\Processor\Process\PropertyGroupShopware6ExportProcess;
+use Ergonode\ExporterShopware6\Infrastructure\Handler\Export\CustomFieldExportCommandHandler;
+use Ergonode\ExporterShopware6\Infrastructure\Processor\Process\CustomFiledShopware6ExportProcess;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
-class PropertyGroupShopware6ExportCommandHandlerTest extends TestCase
+class CustomFieldExportCommandHandlerTest extends TestCase
 {
     /**
      * @var ExportRepositoryInterface|MockObject
@@ -38,9 +38,9 @@ class PropertyGroupShopware6ExportCommandHandlerTest extends TestCase
     private AttributeRepositoryInterface $attributeRepository;
 
     /**
-     * @var PropertyGroupShopware6ExportProcess|MockObject
+     * @var CustomFiledShopware6ExportProcess|MockObject
      */
-    private PropertyGroupShopware6ExportProcess $process;
+    private CustomFiledShopware6ExportProcess $process;
 
     protected function setUp(): void
     {
@@ -59,15 +59,15 @@ class PropertyGroupShopware6ExportCommandHandlerTest extends TestCase
             ->willReturn($this->createMock(AbstractAttribute::class));
         $this->attributeRepository->expects(self::once())->method('load');
 
-        $this->process = $this->createMock(PropertyGroupShopware6ExportProcess::class);
+        $this->process = $this->createMock(CustomFiledShopware6ExportProcess::class);
         $this->process->expects(self::once())->method('process');
     }
 
     public function testHandling(): void
     {
-        $command = $this->createMock(PropertyGroupShopware6ExportCommand::class);
+        $command = $this->createMock(CustomFieldExportCommand::class);
 
-        $handler = new PropertyGroupShopware6ExportCommandHandler(
+        $handler = new CustomFieldExportCommandHandler(
             $this->exportRepository,
             $this->channelRepository,
             $this->attributeRepository,
