+++ module/exporter-shopware-6/tests/Infrastructure/Handler/Export/ProductCrossSellingExportCommandHandlerTest.php
@@ -11,14 +11,16 @@ namespace Ergonode\ExporterShopware6\Tests\Infrastructure\Handler\Export;
 use Ergonode\Channel\Domain\Repository\ChannelRepositoryInterface;
 use Ergonode\Exporter\Domain\Entity\Export;
 use Ergonode\Exporter\Domain\Repository\ExportRepositoryInterface;
-use Ergonode\ExporterShopware6\Domain\Command\Export\CategoryRemoveShopware6ExportCommand;
+use Ergonode\ExporterShopware6\Domain\Command\Export\ProductCrossSellingExportCommand;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
-use Ergonode\ExporterShopware6\Infrastructure\Handler\Export\CategoryRemoveShopware6ExportCommandHandler;
-use Ergonode\ExporterShopware6\Infrastructure\Processor\Process\CategoryRemoveShopware6ExportProcess;
+use Ergonode\ExporterShopware6\Infrastructure\Handler\Export\ProductCrossSellingExportCommandHandler;
+use Ergonode\ExporterShopware6\Infrastructure\Processor\Process\ProductCrossSellingExportProcess;
+use Ergonode\ProductCollection\Domain\Entity\ProductCollection;
+use Ergonode\ProductCollection\Domain\Repository\ProductCollectionRepositoryInterface;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
-class CategoryRemoveShopware6ExportCommandHandlerTest extends TestCase
+class ProductCrossSellingExportCommandHandlerTest extends TestCase
 {
     /**
      * @var ExportRepositoryInterface|MockObject
@@ -31,9 +33,14 @@ class CategoryRemoveShopware6ExportCommandHandlerTest extends TestCase
     private ChannelRepositoryInterface $channelRepository;
 
     /**
-     * @var CategoryRemoveShopware6ExportProcess|MockObject
+     * @var ProductCollectionRepositoryInterface|MockObject
      */
-    private CategoryRemoveShopware6ExportProcess $process;
+    private ProductCollectionRepositoryInterface $productCollectionRepository;
+
+    /**
+     * @var ProductCrossSellingExportProcess|MockObject
+     */
+    private ProductCrossSellingExportProcess $process;
 
     protected function setUp(): void
     {
@@ -47,17 +54,24 @@ class CategoryRemoveShopware6ExportCommandHandlerTest extends TestCase
             ->willReturn($this->createMock(Shopware6Channel::class));
         $this->channelRepository->expects(self::once())->method('load');
 
-        $this->process = $this->createMock(CategoryRemoveShopware6ExportProcess::class);
+        $this->productCollectionRepository = $this->createMock(ProductCollectionRepositoryInterface::class);
+        $this->productCollectionRepository->method('load')
+            ->willReturn($this->createMock(ProductCollection::class));
+        $this->productCollectionRepository->expects(self::once())->method('load');
+
+        $this->process = $this->createMock(ProductCrossSellingExportProcess::class);
         $this->process->expects(self::once())->method('process');
+
     }
 
     public function testHandling(): void
     {
-        $command = $this->createMock(CategoryRemoveShopware6ExportCommand::class);
+        $command = $this->createMock(ProductCrossSellingExportCommand::class);
 
-        $handler = new CategoryRemoveShopware6ExportCommandHandler(
+        $handler = new ProductCrossSellingExportCommandHandler(
             $this->exportRepository,
             $this->channelRepository,
+            $this->productCollectionRepository,
             $this->process
         );
         $handler->__invoke($command);
