+++ module/exporter-shopware-6/tests/Infrastructure/Handler/Export/PropertyGroupExportCommandHandlerTest.php
@@ -1,0 +1,78 @@
+<?php
+/*
+ * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
+ * See LICENSE.txt for license details.
+ */
+
+declare(strict_types=1);
+
+namespace Ergonode\ExporterShopware6\Tests\Infrastructure\Handler\Export;
+
+use Ergonode\Attribute\Domain\Entity\AbstractAttribute;
+use Ergonode\Attribute\Domain\Repository\AttributeRepositoryInterface;
+use Ergonode\Channel\Domain\Repository\ChannelRepositoryInterface;
+use Ergonode\Exporter\Domain\Entity\Export;
+use Ergonode\Exporter\Domain\Repository\ExportRepositoryInterface;
+use Ergonode\ExporterShopware6\Domain\Command\Export\PropertyGroupExportCommand;
+use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
+use Ergonode\ExporterShopware6\Infrastructure\Handler\Export\PropertyGroupExportCommandHandler;
+use Ergonode\ExporterShopware6\Infrastructure\Processor\Process\PropertyGroupShopware6ExportProcess;
+use PHPUnit\Framework\MockObject\MockObject;
+use PHPUnit\Framework\TestCase;
+
+class PropertyGroupExportCommandHandlerTest extends TestCase
+{
+    /**
+     * @var ExportRepositoryInterface|MockObject
+     */
+    private ExportRepositoryInterface $exportRepository;
+
+    /**
+     * @var ChannelRepositoryInterface|MockObject
+     */
+    private ChannelRepositoryInterface $channelRepository;
+
+    /**
+     * @var AttributeRepositoryInterface|MockObject
+     */
+    private AttributeRepositoryInterface $attributeRepository;
+
+    /**
+     * @var PropertyGroupShopware6ExportProcess|MockObject
+     */
+    private PropertyGroupShopware6ExportProcess $process;
+
+    protected function setUp(): void
+    {
+        $this->exportRepository = $this->createMock(ExportRepositoryInterface::class);
+        $this->exportRepository->method('load')
+            ->willReturn($this->createMock(Export::class));
+        $this->exportRepository->expects(self::once())->method('load');
+
+        $this->channelRepository = $this->createMock(ChannelRepositoryInterface::class);
+        $this->channelRepository->method('load')
+            ->willReturn($this->createMock(Shopware6Channel::class));
+        $this->channelRepository->expects(self::once())->method('load');
+
+        $this->attributeRepository = $this->createMock(AttributeRepositoryInterface::class);
+        $this->attributeRepository->method('load')
+            ->willReturn($this->createMock(AbstractAttribute::class));
+        $this->attributeRepository->expects(self::once())->method('load');
+
+        $this->process = $this->createMock(PropertyGroupShopware6ExportProcess::class);
+        $this->process->expects(self::once())->method('process');
+    }
+
+    public function testHandling(): void
+    {
+        $command = $this->createMock(PropertyGroupExportCommand::class);
+
+        $handler = new PropertyGroupExportCommandHandler(
+            $this->exportRepository,
+            $this->channelRepository,
+            $this->attributeRepository,
+            $this->process
+        );
+        $handler->__invoke($command);
+    }
+}
